// FE/typings.d.ts
import { AsteroidCard } from '@/components/AsteroidCard'
export type Asteroid = {
  id: string
  neo_reference_id: string
  name: string
  nasa_jpl_url: string
  absolute_magnitude_h: number
  estimated_diameter: {
    kilometers: {
      estimated_diameter_min: number
      estimated_diameter_max: number
    }
  }
}